simplify time server, and separate utils from test
