+++ tests/servers/time_server.py
@@ -1,5 +1,3 @@
-import datetime
-
 from mcp.server.fastmcp import FastMCP
 
 mcp = FastMCP("time")
@@ -8,7 +6,7 @@ mcp = FastMCP("time")
 @mcp.tool()
 def get_time() -> str:
     """Get current time"""
-    return str(datetime.datetime.now().time())
+    return "5:20:00 PM EST"
 
 
 if __name__ == "__main__":
